pyproject.toml
[tool.poetry]
name = "msr"
version = "0.1.0"
description = ""
authors = ["shate <tomashawro.kontakt@gmail.com>"]
readme = "README.md"
packages = [
    { include = "msr" },
]


[tool.poetry.dependencies]
python = "^3.8"
jupyterlab = "^3.5.0"
scikit-learn = "^1.1.3"
matplotlib = "<3.6.1"
numpy = "^1.23.4"
plotly = "^5.11.0"
pandas = "^1.5.1"
wfdb = "^4.0.0"
tqdm = "^4.64.1"
hydra-colorlog = "^1.2.0"
seaborn = "^0.12.1"
torch = "^1.13.0"
statsmodels = "^0.13.4"
neurokit2 = "^0.2.1"
biosppy = "^1.0.0"
ipywidgets = "7.0"
wget = "^3.2"
mne = "^1.2.1"
nptyping = "^2.3.1"
rich = "^12.6.0"
pytorch-lightning = "^1.8.0.post1"
torchmetrics = "^0.10.2"
sorcery = "^0.2.2"
wandb = "^0.13.5"
pyrootutils = "^1.0.4"
lightgbm = "^3.3.3"
pywavelets = "^1.4.1"
torchinfo = "^1.7.1"
torchvision = "^0.14.1"
fastai = "1.0.61"
thop = "^0.1.1.post2209072238"
mat73 = "^0.59"
hurry-filesize = "^0.9"

[tool.black]
experimental-string-processing = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
